Save Draft in Activity `onStop` instead of Fragment one
